Migrate test_mc_dbk_kern into test_dbk_kernels.py
